requirements.txt
asgiref==3.5.2
Babel==2.11.0
backports.zoneinfo==0.2.1
Django==4.0.8
django-environ==0.9.0
django-storages==1.13.1
docopt==0.6.2
docxcompose==1.3.7
docxtpl==0.16.4
et-xmlfile==1.1.0
gunicorn==20.1.0
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
mssql-django==1.1.3
mysql-connector-python==8.0.31
mysqlclient==2.1.1
num2words==0.5.12
openpyxl==3.0.10
Pillow==9.3.0
pkg_resources==0.0.0
protobuf==3.20.1
pyodbc==4.0.35
python-docx==0.8.11
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
whitenoise==6.2.0
